Configure stream behaviour on failure
By default it will halt since it's likely that the input is corrupted
and the parser it's an unrecoverable state. With `[continue_on_failure:
true]` the failure will be passed down in the pipeline and the parser
will be restarted with the remaining input. With `[raise_on_failure:
true]` a `RuntimeError` will be raised