Stream could return more success per chunk
Drastically changed the stream behaviour and the interaction with the
parser process. There was many problems but the biggest was the
impossibility of returning more than one result per chunk

Let's say we have a parser lit("aa") and upstream is giving us the chunk
"aaaa" we expect the parser to give ["aa", "aa"] to downstream. That was
not possible because the parser was spawned one time per chunk and you
can't give something downstream unless upstream gave you something

Now we spawn the parser only one time (good), we collect each success
sending message from the parser to the stream to flush them downstream
when the parser stops for more input or on failure